requirements.txt
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
google-api-core==2.14.0
google-auth==2.24.0
google-cloud-aiplatform==1.36.4
google-cloud-bigquery==3.13.0
google-cloud-core==2.3.3
google-cloud-resource-manager==1.10.4
google-cloud-storage==2.13.0
google-cloud-vision==3.4.5
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpc-google-iam-v1==0.12.7
grpcio==1.59.3
grpcio-status==1.59.3
idna==3.6
install==1.3.5
numpy==1.26.2
packaging==23.2
pdf2image==1.16.3
Pillow==10.1.0
proto-plus==1.22.3
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-poppler==0.4.1
requests==2.31.0
rsa==4.9
shapely==2.0.2
six==1.16.0
urllib3==2.1.0
